lib <sys/types.h> included in FT_PRINTF_H, type lon long int changed to ssize_t in f_llitoa_base. f_print_int, f_print_ptr, f_print_uns_hex refactored

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: awerebea <awerebea@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/23 16:26:06 by awerebea          #+#    #+#             */
-/*   Updated: 2020/06/05 14:06:37 by awerebea         ###   ########.fr       */
+/*   Updated: 2020/06/05 18:09:16 by awerebea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 # define FT_PRINTF_H
 # include <stdarg.h>
 # include <wchar.h>
+# include <sys/types.h>
 
 typedef struct	s_opts
 {
@@ -32,7 +33,7 @@ char			*ft_printf_strdup(const char *s);
 int				ft_printf_toupper(int c);
 char			*f_strupper(char *s);
 size_t			ft_printf_strlen(const char *s);
-char			*f_llitoa_base(long long n, int base);
+char			*f_llitoa_base(ssize_t n, int base);
 char			*f_ullitoa_base(unsigned long long n, int base);
 int				f_isspec(const char *format, int *i, t_opts *opts);
 int				f_pars_format(va_list ap, const char *format);
